Create README.md
Hello there! I'm Ronald Broden, and I'd like to take you on a journey through my life in the world of software development.

I was born in 1963 in the vibrant city of Westchester, New York, but my roots trace back to Mexico City. Despite my diverse background, it was the picturesque town of Bedford, right in the heart of Westchester County, that truly shaped my formative years.